Isolating CMFW code...
