+++ Opel-Manager/AndroidManifest.xml
@@ -62,7 +62,7 @@
             android:enabled="true"
             android:exported="true" />
 
-        <activity android:name=".view.DeviceListActivity"
+        <activity android:name="com.opel.cmfw.view.BluetoothDeviceSettingActivity"
                   android:theme="@android:style/Theme.DeviceDefault.Dialog"></activity>
     </application>
