+++ Opel-Manager/res/layout/activity_bluetooth_device_setting.xml
