--- Opel-Manager/res/menu/main.xml
@@ -1,16 +1,0 @@
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.opel.opel_manager.view.MainActivity">
-
-    <item
-        android:id="@+id/action_settings"
-        android:orderInCategory="100"
-        android:showAsAction="never"
-        android:title="@string/action_settings"/>
-   
-    <item android:id="@+id/action_config"
-        android:icon="@drawable/ic_launcher"
-        android:showAsAction="always"
-        android:title="config"/>
-  
-</menu>
