+++ Opel-Manager/res/menu/menu_remote_config_ui.xml
