+++ Opel-Manager/res/values/strings.xml
@@ -11,7 +11,7 @@
     <string name="action_event_clear">Clear</string>
     <string name="title_activity_splash_viewer">splashViewer</string>
 
-    <!--  DeviceListActivity -->
+    <!--  CommInitActivity -->
     <string name="scanning">scanning for devices...</string>
     <string name="select_device">select a device to connect</string>
     <string name="none_paired">No devices have been paired</string>
