+++ Opel-Manager/src/com/opel/cmfw/controller/CommControllerListener.java
@@ -1,8 +1,8 @@
-package com.opel.opel_manager.controller.selectiveconnection;
+package com.opel.cmfw.controller;
 
 /**
  * Created by redcarrottt on 2016. 10. 3..
  */
-public interface WifiDirectStateListener {
+public interface CommControllerListener {
     public void onWifiDirectStateChanged(boolean isOn);
 }
