+++ Opel-Manager/src/com/opel/opel_manager/model/OPELAppList.java
@@ -1,5 +1,10 @@
 package com.opel.opel_manager.model;
 
+import android.content.res.Resources;
+import android.graphics.BitmapFactory;
+
+import com.opel.opel_manager.R;
+
 import java.util.ArrayList;
 
 /**
@@ -7,35 +12,52 @@ import java.util.ArrayList;
  */
 
 public class OPELAppList {
-
-    ArrayList<OPELApplication> mList;
+    private ArrayList<OPELApplication> mList;
+    private boolean mIsNativeAppsAdded;
 
     public OPELAppList() {
         mList = new ArrayList<OPELApplication>();
     }
 
+    public void addNativeAppList(Resources appResources) {
+        if (this.mIsNativeAppsAdded == false) {
+            this.add(new OPELApplication("-1", "Camera", BitmapFactory
+                    .decodeResource(appResources, R.drawable.cam), -1));
+            this.add(new OPELApplication("-1", "Sensor", BitmapFactory
+                    .decodeResource(appResources, R.drawable.sensor), -1));
+            this.add(new OPELApplication("-1", "App Market", BitmapFactory
+                    .decodeResource(appResources, R.drawable.market), -1));
+            this.add(new OPELApplication("-1", "App Manager", BitmapFactory
+                    .decodeResource(appResources, R.drawable
+                            .icon_app_manager), -1));
+            this.add(new OPELApplication("-1", "File Manager", BitmapFactory
+                    .decodeResource(appResources, R.drawable.filemanager), -1));
+            this.add(new OPELApplication("-1", "Event Logger", BitmapFactory
+                    .decodeResource(appResources, R.drawable.eventlogger), -1));
+            this.add(new OPELApplication("-1", "Connect", BitmapFactory
+                    .decodeResource(appResources, R.drawable.connect), -1));
+            this.mIsNativeAppsAdded = true;
+        }
+    }
+
     // -> ready
     public void installApplication(OPELApplication app) {
-
-        add(app);
+        this.add(app);
         app.setTypeToInstalled();
     }
 
     // ready -> remove
     public void uninstallApplication(OPELApplication app) {
-
         removeAllApplicationList(app);
-
     }
 
-    public void add(OPELApplication icon) {
-        mList.add(icon);
+    public void add(OPELApplication app) {
+        mList.add(app);
     }
 
-    public void removeAllApplicationList(OPELApplication icon) {
-        for(int i = 0; i< mList.size(); i++){
-            if (mList.get(i).getAppId()
-                    == icon.getAppId()){
+    public void removeAllApplicationList(OPELApplication app) {
+        for (int i = 0; i < mList.size(); i++) {
+            if (mList.get(i).getAppId() == app.getAppId()) {
                 mList.remove(i);
             }
         }
@@ -45,16 +67,13 @@ public class OPELAppList {
         return this.mList;
     }
 
-    public OPELApplication getAppInAllList(String appId) {
-
+    public OPELApplication getApp(String appId) {
         for (int i = 0; i < mList.size(); i++) {
             OPELApplication tmpApp = mList.get(i);
             if (tmpApp.getAppId() == Integer.parseInt(appId)) {
                 return tmpApp;
             }
         }
-
         return null;
     }
-
 }
